Implement Granger-causality and instantaneous causality for deterministic="nc"
Note that test cases for instantaneous causality fail. They would pass with
"self.p+1" instead of "self.p" at the beginning of the method body. I think this might be a bug in JMulTi, since Lutkepohl (2005) is explicitly advising to use a VAR(p)-model and not a VAR(p+1)-model on page 320 (chapter 7.6.4).