add plotting functionality, start implementing test for normality
Plotting functions allow plotting:
- the input data
- forecasts
- impulse responses
When implementing the test for normality I saw that the current implementation in VAR is not producing the same results as JMulTi. For now, I have rewritten the code such that it produces the same results as JMulTi for VARs. The corresponding code for VECMs is not finished yet.