VAR: test and adapt lag order selection
The adaption (line 728 in var_model.py) was made to avoid an exception because the current code isn't able to handle an empty model (no deterministic terms and no lag). Now the program doesn't consider a lag of 0 when choosing the best possible lag in the case of deterministic="nc".
Todo: 1.) Allow for an empty model.  2.) Then include lag=0 in lag order selection also for the case deterministic="nc" (i.e. change the line of code mentioned above).